Extract interface from org.hibernate.boot.model.relational.Database and org.hibernate.boot.model.relational.Schema.
